auto-video-uploader/requirements.txt
Flask==2.2.5
google-auth==2.17.3
google-auth-oauthlib==0.7.1
google-api-python-client==2.86.0
requests==2.31.0
mediapipe
opencv-python
